change value to controlProperty

+++ JellySlider+BEKApps/JellySlider+Rx.swift
@@ -1,6 +1,6 @@
 //
 //  JellySlider+Rx.swift
-//  Playor
+//  BEKApps.com
 //
 //  Created by Behrad Kazemi on 9/14/19.
 //  Copyright © 2019 Behrad Kazemi. All rights reserved.
@@ -9,9 +9,11 @@
 import Foundation
 import RxCocoa
 import RxSwift
+
 extension Reactive where Base: JellySlider {
+	
 	/// Reactive wrapper for `value` property.
-	public var value: ControlProperty<CGFloat> {
+	public var value: ControlProperty<Float> {
 		return controlProperty(
 			editingEvents: [.allEditingEvents, .valueChanged],
 			getter: { slider in
